// src/types/beats/aptos.ts
// Auto-generated by scripts/generate-block-types.ts
// Generated: 2026-03-17
// Sources:
//   https://jiti.indexing.co/networks/aptos/666725417
//   https://jiti.indexing.co/networks/aptos/666725317
//   https://jiti.indexing.co/networks/aptos/666724417
// Samples: 3
export interface AptosBlockTransactionsChanges {
  address: string;
  state_key_hash: string;
  data: Record<string, unknown> | null;
  type: string;
  handle: string;
  key: string;
  value: string;
}

export interface AptosBlockTransactionsEventsGuid {
  creation_number: string;
  account_address: string;
}

export interface AptosBlockTransactionsEventsData {
  epoch?: string;
  failed_proposer_indices?: unknown[];
  hash?: string;
  height?: string;
  previous_block_votes_bitvec?: string;
  proposer?: string;
  round?: string;
  time_microseconds?: string;
  __variant__?: string;
  ask_prices?: string[];
  ask_sizes?: string[];
  bid_prices?: string[];
  bid_sizes?: string[];
  cancelled_ask_prices?: unknown[];
  cancelled_ask_sizes?: unknown[];
  cancelled_bid_prices?: unknown[];
  cancelled_bid_sizes?: unknown[];
  market?: string;
  order_id?: string;
  parent?: string;
  previous_seq_num?: string;
  sequence_number?: string;
  user?: string;
  execution_gas_units?: string;
  io_gas_units?: string;
  storage_fee_octas?: string;
  storage_fee_refund_octas?: string;
  total_charge_gas_units?: string;
}

export interface AptosBlockTransactionsEvents {
  guid: AptosBlockTransactionsEventsGuid;
  sequence_number: string;
  type: string;
  data: AptosBlockTransactionsEventsData;
}

export interface AptosBlockTransactionsBlockMetadataExtension {
  randomness: null;
  type: string;
}

export interface AptosBlockTransactionsBlockEndInfo {
  block_gas_limit_reached: boolean;
  block_output_limit_reached: boolean;
  block_effective_block_gas_units: number;
  block_approx_output_size: number;
}

export interface AptosBlockTransactionsPayload {
  function: string;
  type_arguments: unknown[];
  arguments: (Record<string, unknown> | string | unknown[])[];
  type: string;
}

export interface AptosBlockTransactionsSignature {
  public_key: string;
  signature: string;
  type: string;
}

export interface AptosBlockTransactions {
  version: string;
  hash: string;
  state_change_hash: string;
  event_root_hash: string;
  state_checkpoint_hash: null | string;
  gas_used: string;
  success: boolean;
  vm_status: string;
  accumulator_root_hash: string;
  changes: AptosBlockTransactionsChanges[];
  id: string;
  epoch: string;
  round: string;
  events: AptosBlockTransactionsEvents[];
  previous_block_votes_bitvec: number[];
  proposer: string;
  failed_proposer_indices: unknown[];
  timestamp: string;
  block_metadata_extension: AptosBlockTransactionsBlockMetadataExtension;
  type: string;
  block_end_info: AptosBlockTransactionsBlockEndInfo;
  sender?: string;
  sequence_number?: string;
  max_gas_amount?: string;
  gas_unit_price?: string;
  expiration_timestamp_secs?: string;
  payload?: AptosBlockTransactionsPayload;
  signature?: AptosBlockTransactionsSignature;
  replay_protection_nonce?: string;
}

export interface AptosBlock {
  block_height: string;
  block_hash: string;
  block_timestamp: string;
  first_version: string;
  last_version: string;
  transactions: AptosBlockTransactions[];
  _network: string;
  _vm: string;
}